+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.12.3@DEV
+  * Fixed bug with naming of validator in errors array. Removed suffix 'validator'.
+Old validator name is kept for back compatibility and will be removed in next versions
+
 ## 1.12.2(2015-01-13)
   * `Operator` refactoring and bugfixes
   * Fix Collection::update when update data set as array